refactor CondaNoticeButton
